[xlib: remove output vs. crtcs check]
Remove the equality check for the number of outputs and crtcs.

The implementation never relies on the assumption that both counts are
equal. Further, the counts might actually differ without indicating an
error. Hence, keeping the check in place would cause the function to
fail for no reason. In fact, the check caused winfade to abort on a
Thinkpad Z13.