[xdo: fix wrong window position]
issue:

Window position is off by a factor of 2.

reason:

XTranslateCoordinates() assumes coordinates relative to the parent
window. Hence, in order to get the position of a window relative to the
top, left corner of the screen (0, 0) have to be used as the coordinates
to translate, rather than the coordinates returned by
XGetWindowAttributes().

fix:

Use (0, 0) as input coordinates to XTranslateCoordinates().